Organize into file AudioCustomProcessingDelegate
